Share Deno CLI helpers
